Moved extension to it's own group.
