+++ Mango (Social Manga)/Controller/TitleSearch.swift
@@ -50,37 +50,5 @@ class TitleSearch: UIViewController, UITableViewDelegate, UITableViewDataSource 
     }
 }
 
-//MARK: - Extensions
-extension TitleSearch: UISearchBarDelegate {
-    
-    func searchBarTextDidBeginEditing(_ searchBar: UISearchBar) {
-        isSearching = true
-    }
-    
-    func searchBarTextDidEndEditing(_ searchBar: UISearchBar) {
-        isSearching = false
-    }
-    
-    func searchBarCancelButtonClicked(_ searchBar: UISearchBar) {
-        isSearching = false
-    }
-    
-    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
-        self.searchBar.resignFirstResponder()
-        searchFilter(searchBar)
-        table?.reloadData()
-    }
-    
-    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-    }
-    
-    fileprivate func searchFilter(_ searchBar: UISearchBar) {
-        filteredArray.removeAll(keepingCapacity: false)
-        let predicateString = searchBar.text!
-        filteredArray = testArray.filter( {$0.range(of: predicateString) != nil} )
-        filteredArray.sort {$0 < $1}
-        isSearching = (filteredArray.count == 0) ? false: true
-    }
-    
-}
+
